Allow looking ahead to see if a cell leads to an unwinnable situation
Limit how far in advance we look ahead, to avoid a brute force solution that cannot be explained.

Still need to figure out how to explain the limited look ahead presented here.
May need to add another property to Move that includes the chain of moves that led to
the unwinnable situation.